# using ImageBase.fdiff and some other pre-allocation to improve the performance