--- a/Project.toml
+++ b/Project.toml
@@ -6,12 +6,11 @@ version = "0.1.0"
 [deps]
 FFTW = "7a1cc6ca-52ef-59f5-83cd-3a7055c09341"
 ImageBase = "c817782e-172a-44cc-b673-b171935fbb9e"
-ImageCore = "a09fc81d-aa75-5fe9-8630-4744c3626534"
 ImageFiltering = "6a3955dd-da59-5b1f-98d4-e7296123deb5"
 
 [compat]
 FFTW = "1"
-ImageCore = "0.8, 0.9"
+ImageBase = "0.1.2"
 ImageFiltering = "0.6"
 julia = "1"