--- a/docs/src/reference.md
+++ b/docs/src/reference.md
@@ -12,13 +12,6 @@ smooth
 smooth!
 ```
 
-```@docs
-forwarddiff
-forwarddiff!
-backdiff
-backdiff!
-```
-
 ## Algorithm
 
 ```@docs