--- a/src/ImageSmooth.jl
+++ b/src/ImageSmooth.jl
@@ -3,6 +3,7 @@ module ImageSmooth
 using ImageFiltering, FFTW
 using ImageCore
 using ImageCore: GenericGrayImage, GenericImage
+using ImageBase
 
 #TODO: port SmoothAPI to ImagesAPI
 include("SmoothAPI/SmoothAPI.jl")
@@ -19,9 +20,6 @@ include("algorithms/l0_smooth.jl") # L0Smooth
 export
     # generic API
     smooth, smooth!,
-
-    # forward difference and backward difference
-    forwarddiff, forwarddiff!, backdiff, backdiff!,
 
     # Algorithms
     L0Smooth