--- a/src/ImageSmooth.jl
+++ b/src/ImageSmooth.jl
@@ -9,9 +9,6 @@ include("SmoothAPI/SmoothAPI.jl")
 import .SmoothAPI: AbstractImageSmoothAlgorithm,
                    smooth, smooth!
 
-include("utils.jl")
-include("compat.jl")
-
 # Smooth algorithms
 
 include("algorithms/l0_smooth.jl") # L0Smooth