--- a/src/algorithms/l0_smooth.jl
+++ b/src/algorithms/l0_smooth.jl
@@ -131,8 +131,8 @@ function (f::L0Smooth)(out::AbstractArray{<: Number},
     𝛥₂𝑆 = similar(𝑆)
     s³ = similar(𝑆)
     s¹ = zeros(1, N, M)
-    𝛻₁ℎ = similar(𝑆)
-    𝛻₂𝑣 = similar(𝑆)
+    𝛥₁ᵀℎ = similar(𝑆)
+    𝛥₂ᵀ𝑣 = similar(𝑆)
 
     Normin = similar(ℱ𝐼) 
     t¹ = trues(1, N, M)
@@ -167,16 +167,17 @@ function (f::L0Smooth)(out::AbstractArray{<: Number},
         # For equation (8), ℎ = 𝛥₁𝑆, 𝑣 = 𝛥₂𝑆
         # According to Convolution Theorem, ℱ(𝑓₁ * 𝑓₂) = ℱ(𝑓₁) ⋅ ℱ(𝑓₂)
         # ℱ is the FFT operator, * is a convolution operator, ⋅ is a matrix dot times operator
-        # We can compute ℱ(∂₁)* ⋅ ℱ(ℎ) and ℱ(∂₂)* ⋅ ℱ(𝑣) by computing ℱ(𝛻₁ℎ) and ℱ(𝛻₂𝑣)
+        # We can compute ℱ(∂₁)* ⋅ ℱ(ℎ) and ℱ(∂₂)* ⋅ ℱ(𝑣) by computing ℱ(𝛥₁ᵀℎ) and ℱ(𝛥₂ᵀ𝑣)
         # ∂₁ and ∂₂ are the difference operators along horizontal axis and vertical axis, respectivly
-        # 𝛻₁() and 𝛻₂() indicate the backward difference along horizontal axis and vertical axis
-        fdiff!(𝛻₁ℎ, 𝛥₁𝑆, dims = 3, rev=true, boundary=:periodic)
-        fdiff!(𝛻₂𝑣, 𝛥₂𝑆, dims = 2, rev=true, boundary=:periodic)
-        @. 𝛻₁ℎ = -𝛻₁ℎ
-        @. 𝛻₂𝑣 = -𝛻₂𝑣
+        # ℱ()* denotes the complex conjugate
+        # 𝛥₁ᵀ() and 𝛥₂ᵀ() indicate the backward difference along horizontal axis and vertical axis
+        fdiff!(𝛥₁ᵀℎ, 𝛥₁𝑆, dims = 3, rev=true, boundary=:periodic)
+        fdiff!(𝛥₂ᵀ𝑣, 𝛥₂𝑆, dims = 2, rev=true, boundary=:periodic)
+        @. 𝛥₁ᵀℎ = -𝛥₁ᵀℎ
+        @. 𝛥₂ᵀ𝑣 = -𝛥₂ᵀ𝑣
 
         # Computing S via equation (8)
-        @. Normin = complex(𝛻₁ℎ + 𝛻₂𝑣)
+        @. Normin = complex(𝛥₁ᵀℎ + 𝛥₂ᵀ𝑣)
         fft!(Normin, (2, 3))
         @. ℱ𝑆 = (ℱ𝐼 + 𝛽 * Normin) / (1 + 𝛽 * Denormin)
         ifft!(ℱ𝑆, (2, 3))