### Some improvement
* **`ImageBase.fdiff`** is used to improve the performance, solves https://github.com/johnnychen94/ImageSmooth.jl/issues/4#issue-963142327
* **Excluding some unnecessary allocation** from codes like `similar(float64.(input))` or `sum()`, solves https://github.com/johnnychen94/ImageSmooth.jl/issues/3#issue-963142254

Previous performance is as follow:
```julia
julia> img_gray = testimage("cameraman");

julia> img_rgb = testimage("lena_color_512");

julia> fₛ = L0Smooth()
L0Smooth(0.02, 2.0, 100000.0)

julia> @btime smooth(img_gray, fₛ)
  418.836 ms (1712 allocations: 316.75 MiB)
julia> @btime smooth(img_rgb, fₛ)
  1.155 s (1934 allocations: 1.03 GiB)
```

Current performance is as follow:
```julia
julia> img_gray = testimage("cameraman");

julia> img_rgb = testimage("lena_color_512");

julia> f = L0Smooth()
L0Smooth(0.02, 2.0, 100000.0)

julia> @btime smooth(img_gray, f)
  393.819 ms (1449 allocations: 138.77 MiB)

julia> @btime smooth(img_rgb, f)
  1.013 s (1627 allocations: 387.81 MiB)
```
We can see that the **allocations are extremely reduced**, however, the **improvement of running time is not as remarkable as the allocations**. 

Maybe we can do some **further improvement**, or the **main time-consuming factor** is the process of `fft!` and `ifft!` so that the performance of this algorithm stay at this level.